// types/friendshipLevel.ts
type FriendshipLevel = 
  | "BFF" 
  | "Best Friend" 
  | "Inseparable" 
  | "Close Friend" 
  | "Great Friend" 
  | "Good Friend" 
  | "Friend" 
  | "Buddy" 
  | "Pal" 
  | "Acquaintance" 
  | "Opponent" 
  | "Rival" 
  | "Foe" 
  | "Adversary" 
  | "Antagonist" 
  | "Enemy" 
  | "Villain" 
  | "Malefactor" 
  | "Nemesis" 
  | "Arch Nemesis";

  export default FriendshipLevel;